% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{abondanceScore}
\alias{abondanceScore}
\title{abondanceScore}
\usage{
abondanceScore(vect, th = 0)
}
\arguments{
\item{vect}{: a numerical vector}

\item{th}{: the threshold to be applied, default is 0}
}
\value{
a mean abundance
}
\description{
Computes the sum of the vectors divided by the prevalence
}
\details{
\code{abondanceScore}
}
\author{
Edi Prifti
}
