% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querries.R
\name{aggregateProfiles}
\alias{aggregateProfiles}
\title{aggregateProfiles}
\usage{
aggregateProfiles(list.profiles, max.size = 25, min.size = max.size)
}
\arguments{
\item{list.profiles}{: list of matrix profiles}

\item{max.size}{: the maximum number of rows to be selected in the final agregated matrix.
default   max.size=25.}

\item{min.size}{: this is the minimum number of rows rows to be selected in the final aggregated matrix. 
If a group has less it will be discarded. By default min.size=max.size.}
}
\value{
an aggregated profile matrix.
}
\description{
This function takes a list of profile matrixes and returns an aggregated big matrix.
The individual matrixes can be filtered in size so that the first X rows are used for each of them.
This function is used to prepare the data and plot different MGS as barcodes.
}
\details{
\code{aggregateProfiles}
}
\author{
Emmanuelle Le Chatelier
}

