% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{computeSignalMetrics}
\alias{computeSignalMetrics}
\title{computeSignalMetrics}
\usage{
computeSignalMetrics(dat)
}
\arguments{
\item{dat}{: a matrix where operations will be performed on the columns. Please transpose if operations
are needed in the rows.}
}
\value{
a matrix of results where scores are in the columns
}
\description{
Computes scores of data variation within a given MGS cluster.
}
\details{
\code{computeSignalMetrics}
}
\author{
Edi Prifti
}
