% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querries.R
\name{connectivity}
\alias{connectivity}
\title{connectivity}
\usage{
connectivity(prof, method = "pearson", th = 0, soft = FALSE)
}
\arguments{
\item{prof}{: a profile matrix. This data is used to compute correlations.}

\item{method}{: the correlation method, default = pearson.}

\item{th}{: default 0, if >0 than this threshold will be applied to compute a hard threholded connectivity.}

\item{soft}{: default = FALSE, if TRUE and when th > 0, the connectivity is computed as the soft threholded, sum of correlations 
above the threshold}
}
\value{
a vector of connectivity
}
\description{
This function computes the intra-row correlation and applied a threshold to compute the connections of each row.
A connectivity vector is returned.
}
\details{
\code{connectivity}
}
\note{
this connectivity does not use a hard thresholding but is based on a total correlation score
}
\author{
Emmanuelle le Chatelier & Edi Prifti
}
