% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsizing.R
\name{countSampledGenesGC}
\alias{countSampledGenesGC}
\title{countSampledGenesGC}
\usage{
countSampledGenesGC(v.samp)
}
\arguments{
\item{v.samp}{: a character vector of the sampled indexed reads (output of sampleRandlomly)}
}
\value{
a table with counts for each gene
}
\description{
counts the number of genes that have been sampled by un_indexing
}
\details{
\code{countSampledGenesGC}
}
\note{
an optimized version for the gene count downsizing
}
\author{
Edi Prifti
}
