% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapreduce.R
\name{deleteData}
\alias{deleteData}
\title{deleteData}
\usage{
deleteData(name = "data_part")
}
\arguments{
\item{name}{: the name of the split files in the disk preceding the incremental number, default="data_part"}
}
\value{
nothing to be returned
}
\description{
This function will delete the original and treated split temporary files to celan the workspace.
}
\details{
\code{deleteData}
}
\author{
Edi Prifti
}

