% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsizing.R
\name{downsizedRichnessL2T}
\alias{downsizedRichnessL2T}
\title{downsizedRichnessL2T}
\usage{
downsizedRichnessL2T(richness.list)
}
\arguments{
\item{richness.list}{: the result of the downsizeGC.all method}
}
\value{
A matrix with the samples in rows and the downsizing in columns
}
\description{
This procedure takes a list that is the result of the downsizeGC.all method and transforms 
it in a matrix of meaned downsied values. Each element of this list contains downsizing results for a given 
sample. This result is a matrix in lines the number of simulations and in columns the different downsizing levels
}
\details{
\code{downsizedRichnessL2T}
}
\author{
Edi Prifti & Emmanuelle Le Chatelier
}

