% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsizing.R
\name{downsizeGC}
\alias{downsizeGC}
\title{downsizeGC}
\usage{
downsizeGC(data, level = 1.1e+07, repetitions = 30, silent = FALSE)
}
\arguments{
\item{data}{: raw read count matrix with gene_ids as rownames}

\item{level}{: default 11e6, the downsizing level number of reads to be selected randomly.}

\item{repetitions}{: default 30, the number of times the drawing is performed. Usually 30 or 10 to speed things out}

\item{silent}{: default is FALSE prints the status of downsizing}
}
\value{
a matrix containing in rows a vector for each repetition and in columns the number of downsized genes for each sample
}
\description{
This function takes a matrix with raw reads counts and computes the number of genes 
at a given downsizing level a given number of times.
}
\details{
\code{downsizeGC}
}
\note{
if the downsizing level is higher than the number of reads for a given sample than the result will be NA
}
\author{
Edi Prifti & Emmanuelle Le Chatelier
}

