% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsizing.R
\name{downsizeMatrix}
\alias{downsizeMatrix}
\title{downsizeMatrix}
\usage{
downsizeMatrix(data, level = 1.1e+07, HQ_reads = NULL, repetitions = 1,
  silent = FALSE)
}
\arguments{
\item{data}{: raw read count matrix with gene_ids as rownames}

\item{level}{: default 11E06, the downsizing levels number of reads to be selected randomly.}

\item{HQ_reads}{: To take into account a vector of high quality reads that are not mapped.}

\item{repetitions}{: default 1, This can be also computed several times, but one is the error minimal downsizing strategy}

\item{silent}{: default is FALSE prints the status of downsizing}
}
\value{
downsized read gene count matrix corresponding to the mean counts obtained with the selected number of independant 
     downsizing procedure
}
\description{
takes a matrix with raw read gene counts and converts it to a downsized matrix with identical number of mapped reads 
     for each sample (column)
}
\details{
\code{downsizeMatrix}
}
\note{
if the downsizing level is higher than the number of reads for a given sample than the result will be NA
}
\author{
Edi Prifti & Emmanuelle Le Chatelier
}

