% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querries.R
\name{extractProfiles}
\alias{extractProfiles}
\title{extractProfiles}
\usage{
extractProfiles(genebag, data, size.max = 15000, size.min = 1,
  silent = TRUE)
}
\arguments{
\item{genebag}{: vector or list of gene_ids}

\item{data}{: raw count or frequency matrix with genes_ids as rawnames}

\item{size.max}{: default 15000, maximum number of profiles to be extracted}

\item{size.min}{: default 1, the minimal size threshold above which a group of genes is selected.
This is used to extract multiple profiles and filtering the list with a minimal number of genes}

\item{silent}{: default TRUE, detailling and following computation progress}
}
\value{
a matrix or a list of profile matrixes
}
\description{
This function extracts the profiles from a gene profile matrix of a group of genes or a list 
of groups of genes. It can also restrict the size of the result
}
\details{
\code{extractProfiles}
}
\author{
Edi Prifti & Emmanuelle Le Chatelier
}

