% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{extractSignificant}
\alias{extractSignificant}
\title{extractSignificant}
\usage{
extractSignificant(relation.matrix, interest, threshold = 0.05)
}
\arguments{
\item{relation.matrix}{: a matrix of p- produced by or q-values by phenoPairwiseRelations()}

\item{interest}{: a vector of variable names of interest.}

\item{threshold}{: default 0.05 needed to select significant relations}
}
\value{
a list of vectors containing p-values or q-values along with the names of the variables.
}
\description{
This function will extract a list of vectors p- or q-values from an object produced by phenoPairwiseRelations.
}
\details{
\code{extractSignificant}
}
\author{
Edi Prifti
}
