% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{filt.hierClust}
\alias{filt.hierClust}
\title{filt.hierClust}
\usage{
filt.hierClust(mat.rho, hclust.method = "ward", margins = c(6, 6),
  side.col.c = NULL, side.col.r = NULL, size = 10, plot = TRUE,
  filt = 0.5)
}
\arguments{
\item{mat.rho}{: square correlation matrix with ids (can be used for also other than just samples)}

\item{hclust.method}{: the hierarchical clustering method, by default it is the ward.D method}

\item{margins}{: change margins of the graph (default = c(6, 6))}

\item{side.col.c}{: a vector of colors to be applied in the columns, usually depincting a class}

\item{side.col.r}{: a vector of colors to be applied in the rows, usually depincting a class}

\item{size}{: the number of samples in the resulting ordered matrix}

\item{plot}{: logical default TRUE. It will plot the heatmap of the similarity with the hierchical clustering}

\item{filt}{: default is 0.5 and is the filtering threshold to be applied}
}
\value{
it will return a matrix with samples in rows and their closely related ones on the columns along with the 
correlation score.
}
\description{
This function takes as input a square similarity matrix and searches for clusters of samples with strong associations
and extracts the sub matrix with the closely related sampless. Only positive correlations are considered here.
}
\details{
\code{filt.hierClust}
}
\author{
Emmanuelle Le Chatelier & Edi Prifti
}

