% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filterListGenebags}
\alias{filterListGenebags}
\title{filterListGenebags}
\usage{
filterListGenebags(list.genebags, size.min = 0, size.max = 15000)
}
\arguments{
\item{list.genebags}{: a list of genebags that needs to be filtered}

\item{size.min}{: the minimal size threshold above which mgs are selected, default 0}

\item{size.max}{: the maximal size threshold above which mgs are selected, default 15000}
}
\value{
a list of selected genebags with their original elements (usually geneids)
}
\description{
This function filters a list of mgs with gene_id >= a given mgs gene number.
}
\details{
\code{filterListGenebags}
}
\note{
This is the former filterListMGS function
}
\author{
Emmanuelle Le Chatelier
}
