% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filterMat}
\alias{filterMat}
\title{filterMat}
\usage{
filterMat(mat, filt = 0)
}
\arguments{
\item{mat}{: matrix of counts}

\item{filt}{: filtering threshold in percentage}
}
\value{
a cleaned matrix
}
\description{
This function filters a matrix (mat) based on the rate of positive value (filt) 
in each individual under a given degree of presence (filt), all sparse values are put to 0.
}
\details{
\code{filterMat}
}
\author{
Edi Prifti & Emmanuelle Le Chatelier
}
