% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{hierClust}
\alias{hierClust}
\title{hierClust}
\usage{
hierClust(data, side = "col", dist = "correlation", cor.type = "spearman",
  hclust.method = "ward.D", side.col.c = NULL, side.col.r = NULL,
  plot = TRUE)
}
\arguments{
\item{data}{: frequency matrix with gene_ids in the rownames}

\item{side}{: the distance can be performed on the columns or on the rows}

\item{dist}{: the type of distance used. By default this is correlation based similarity}

\item{cor.type}{: when correlation matrix, the default is spearman}

\item{hclust.method}{: the hierarchical clustering method, by default it is the ward.D method}

\item{side.col.c}{: a vector of colors to be applied in the columns, usually depincting a class}

\item{side.col.r}{: a vector of colors to be applied in the rows, usually depincting a class}

\item{plot}{: logical default TRUE. It will plot the heatmap of the similarity with the hierchical clustering}
}
\value{
it will return a list of three variables, the correlation matrix, the distance matrix and the hclust object
}
\description{
This function computes the pairwise distance between samples and computes a hierarchical clustering
that is further depicted as a heatmap graphic
}
\details{
\code{hierClust}
}
\note{
updated hierClust functions by elechat april 7th 2015 added options SideColors added + spearman == pearson(rank)
}
\author{
Edi Prifti
}

