% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs_3.3_metahit_genesize}
\alias{hs_3.3_metahit_genesize}
\title{hs_3.3_metahit_genesize}
\description{
The gene size for the hs_3.3_metahit catalogue, needed for the normalization procedure
}
\details{
CATALOG HS_3.3_METAHIT
}
\author{
Edi Prifti & Emmanuelle Le Chatelier
}
\keyword{data}
\keyword{gnesize}
