% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs_3.3_metahit_sample_dat_freq}
\alias{hs_3.3_metahit_sample_dat_freq}
\title{hs_3.3_metahit_sample_dat_freq}
\description{
this dataset contains only the first 100000 genes of the catalogue for size
purposes
}
\author{
Edi Prifti & Emmanuelle Le Chatelier
}
\keyword{dataset}
\keyword{test}
