% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsizing.R
\name{indexReads}
\alias{indexReads}
\title{indexReads}
\usage{
indexReads(v, silent = FALSE)
}
\arguments{
\item{v}{: an integer raw count vector}

\item{silent}{: default FALSE, debugging}
}
\value{
an character indexed vector
}
\description{
for a given column of the raw count matrix this function indexes the count vector 
       in a vector of reads of the length of the sum of the counts. It allows after to randomly 
       select a subset
}
\details{
\code{indexReads}
}
\author{
Edi Prifti
}
