% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapreduce.R
\name{launchTask}
\alias{launchTask}
\title{launchTask}
\usage{
launchTask(input, output, script, nb_proc = NULL, resume = TRUE,
  pattern = "data_part")
}
\arguments{
\item{input}{: a folder containing elements that are to be processed}

\item{output}{: a folder where the processed results will be writen}

\item{script}{: the R script that is to be executed in parallel}

\item{nb_proc}{: number of processes to run the task}

\item{resume}{: resume the computation for the remaining shares (default=TRUE)}

\item{pattern}{: the pattern of the files to be used in the // computation procedure}
}
\value{
nothing
}
\description{
This function distributes the calculations as separate processes in a multi-thread server.
}
\details{
\code{launchTask}
}
\note{
the number of processors may be specified. TODO: add the argument
}
\author{
Edi Prifti
}

