% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{lmp}
\alias{lmp}
\title{lmp}
\usage{
lmp(modelobject)
}
\arguments{
\item{modelobject}{: a linear model object as produced by lm()}
}
\value{
a p-value
}
\description{
This function will extract the p-value from a linear model object. It is used by phenoPairwiseRelations
}
\details{
\code{lmp}
}
\author{
Edi Prifti
}
