% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapreduce.R
\name{mapResults}
\alias{mapResults}
\title{mapResults : merge data from // computing (the reduce process)}
\usage{
mapResults(folder = ".", pattern = "data_part", type = "rows",
  verbose = TRUE)
}
\arguments{
\item{folder}{: the folder where the results are found}

\item{pattern}{: the pattern of the split files in the disk preceding the incremental number, default="data_part"}

\item{type}{: a string c(rows, cols, list) indicating how to merge the data. This should be the same as the one 
used in the splitData procedure.}

\item{verbose}{: print running status information (default = TRUE).}
}
\value{
a merged dataframe : Caution ! The merged results may be shuffeled. It is up to the user to reorder
the data accordingly.
}
\description{
This function loads the results once finished and merges them together in one dataframe. For the moment this is a slow procedure and we are testing ways to accelerate the binding.
}
\details{
\code{mapResults}
}
\author{
Edi Prifti
}

