% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{matrix2Integer}
\alias{matrix2Integer}
\title{matrix2Integer}
\usage{
matrix2Integer(data, verbose = TRUE)
}
\arguments{
\item{data}{: numerical data frame}

\item{verbose}{: print running information (default = TRUE)}
}
\value{
Integer matrix
}
\description{
converts numeric matrix into an integer for memory optimization
}
\details{
\code{matrix2Integer}
}
\author{
Edi Prifti
}
