% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mgs_hs_3.3_metahit_sup500}
\alias{mgs_hs_3.3_metahit_sup500}
\title{mgs_hs_3.3_metahit_sup500}
\description{
This is the mgs catalogue of the 3.3 metahit gene catalogue and contains 778 MGS
of size greater than 500 genes. It has also been manually curated.
}
\author{
Edi Prifti & Emmanuelle Le Chatelier
}
\keyword{catalog}
\keyword{mgs}

