% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normFreqRPKM}
\alias{normFreqRPKM}
\title{normFreqRPKM}
\usage{
normFreqRPKM(dat, cat = NULL, HQ_reads = NULL)
}
\arguments{
\item{dat}{: raw counts data matrix with gene_ids as rownames}

\item{cat}{: the current working catalogue where the reads are mapped and counted, (i.e. hs_3.3_metahit, hs_3.9_metahit)
This can also be a vector of genelength values that correspond to the number of rows in the dat matrix and are 
ordered respectively}

\item{HQ_reads}{: To take into account a vector of high quality reads that are not mapped.}
}
\value{
a normalized frequency matrix
}
\description{
converts a raw count matrix onto a frequency matrix using the RPKM normalization method.
       This method consists of two consecutive steps, first dividing the raw counts by the length of 
       the gene sequence and the second shrinking the signal per column to a sum of 1
}
\details{
\code{normFreqRPKM}
}
\author{
Edi Prifti & Emmanuelle Le Chatelier
}
