% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normFreqTC}
\alias{normFreqTC}
\title{normFreqTC}
\usage{
normFreqTC(dat, HQ_reads = NULL)
}
\arguments{
\item{dat}{: raw counts data matrix with gene_ids as rownames}

\item{HQ_reads}{: To take into account a vector of high quality reads that are not mapped.}
}
\value{
a normalized frequency matrix
}
\description{
converts a raw count matrix onto a frequency matrix using the TC (total count) normalization method.
       This method consists of scaling the signal by the total counts per each sample
}
\details{
\code{normFreqTC}
}
\author{
Edi Prifti
}
