% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{phenoPairwiseRelations}
\alias{phenoPairwiseRelations}
\title{phenoPairwiseRelations}
\usage{
phenoPairwiseRelations(data, adjust = "BH", verbose = FALSE)
}
\arguments{
\item{data}{: bioclinical data with variables on the rows and samples on the columns}

\item{adjust}{: method to adjust for multiple testing (default="BH)}

\item{verbose}{: default=FALSE. If TRUE information will be printed to follow the progression.}
}
\value{
a list of two matrixes containing the p-values and the multiple testing adjustment.
}
\description{
This function will compute all the relations between different variables adapting different statistical tests as a 
function of the data type. It will adjust the p-value matrix for multiple testing.
}
\details{
\code{phenoPairwiseRelations}
}
\author{
Edi Prifti
}

