% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotBarcode}
\alias{plotBarcode}
\title{plotBarcode}
\usage{
plotBarcode(data, main = "")
}
\arguments{
\item{data}{: a frequency matrix to be visualized}

\item{main}{: the main title of the plot empty by default}
}
\value{
nothing
}
\description{
plots the intensity of a frequency matrix with a 4-fold color step
}
\details{
\code{plotBarcode}
}
\note{
this may be slightly affected by the size of the catalogue when comparing different studies
}
\author{
Edi Prifti & Emmanuelle Le Chatelier
}

