% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotBarcode2}
\alias{plotBarcode2}
\title{plotBarcode2}
\usage{
plotBarcode2(data, main = "", ylabl = "", ylabr = "",
  col.axisl = "white", col.axisr = "white", box = FALSE)
}
\arguments{
\item{data}{: a frequency matrix to be visualized}

\item{main}{: the main title of the plot empty by default}

\item{ylabl}{: label for the left y axis}

\item{ylabr}{: label for the right y axis}

\item{col.axisl}{: color for the left y axis}

\item{col.axisr}{: color for the right y axis}

\item{box}{: default FALSE.}
}
\value{
nothing
}
\description{
plots the intensity of a frequency matrix with a 4-fold color step. Usually used
     for complex figures where different MGS are overlapped and annotated with different data.
}
\details{
\code{plotBarcode2}
}
\note{
this may be slightly affected by the size of the catalogue when comparing different studies
}
\author{
Edi Prifti & Emmanuelle Le Chatelier
}
