% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotBarcodeBW}
\alias{plotBarcodeBW}
\title{plotBarcodeBW}
\usage{
plotBarcodeBW(data, main = "")
}
\arguments{
\item{data}{: a frequency matrix to be visualized}

\item{main}{: the main title of the plot empty by default}
}
\value{
nothing
}
\description{
plots in black when a signal is different from zero and white otherwise
}
\details{
\code{plotBarcodeBW}
}
\author{
Edi Prifti
}

