% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotCol}
\alias{plotCol}
\title{plotCol: a matrix of colored cells when debugging}
\usage{
plotCol(col, nrow = 1, ncol = ceiling(length(col)/nrow),
  txt.col = "black")
}
\arguments{
\item{col}{: a vector of colors}

\item{nrow}{: number of rows in of the matrix (1 = default)}

\item{ncol}{: number of columns in of the matrix. If nothing is used it will be computed automatically to fit the cols.}

\item{txt.col}{: font color.}
}
\value{
nothing
}
\description{
a matrix of colored cells when debugging
}
\details{
\code{plotCol}
}
\author{
Edi Prifti
}
