% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotCors}
\alias{plotCors}
\title{plotCors}
\usage{
plotCors(data, main = "")
}
\arguments{
\item{data}{: a frequency matrix to be visualized}

\item{main}{: the main title of the plot empty by default}
}
\value{
nothing
}
\description{
plots a heatmap of a matrix composed of correlation values from -1 to 1. The blue 
     colors are negative correlations while the red are positive
}
\details{
\code{plotCors}
}
\author{
Edi Prifti
}
