% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotCors2}
\alias{plotCors2}
\title{plotCors2}
\usage{
plotCors2(data, main = "", grid = TRUE, axes = TRUE, topdown = TRUE)
}
\arguments{
\item{data}{: a frequency matrix to be visualized}

\item{main}{: the main title of the plot empty by default}

\item{grid}{: add the grid on top of the image (default=TRUE)}

\item{axes}{: plot the axes (default=TRUE)}

\item{topdown}{: reverse order of rows in the matrix (default=TRUE)}
}
\value{
nothing
}
\description{
similar to plotCors but with more levels of colors
}
\details{
\code{plotCors2}
}
\author{
Edi Prifti
}

