% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotMGSQuality}
\alias{plotMGSQuality}
\title{plotMGSQuality}
\usage{
plotMGSQuality(dat, main = "mgs", return.scores = TRUE)
}
\arguments{
\item{dat}{: a matrix where operations will be performed on the columns. Please transpose if operations
are needed in the rows. This is typically an MGS frequency matrix with genes in the rows.}

\item{main}{: the name of the plot}

\item{return.scores}{: weather the function should return the computed scores or not (default=TRUE).}
}
\value{
a matrix of results where scores are in the columns
}
\description{
Visualized scores of data variation within a given MGS cluster as well as the barcode of the MGS.
A subset of 50 most connected genes is also plotted the same way.
}
\details{
\code{plotMGSQuality}
}
\author{
Edi Prifti
}

