% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotPvals}
\alias{plotPvals}
\title{plotPvals}
\usage{
plotPvals(data, main = "")
}
\arguments{
\item{data}{: a frequency matrix to be visualized}

\item{main}{: the main title of the plot empty by default}
}
\value{
nothing
}
\description{
plots a heatmap of a matrix composed of p-values. Gray is not significant at p=0.05
     and significance decrases from skyblue to darkred
}
\details{
\code{plotPvals}
}
\author{
Edi Prifti
}

