% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{presenceScore}
\alias{presenceScore}
\title{presenceScore}
\usage{
presenceScore(vect, th = 0)
}
\arguments{
\item{vect}{: a numerical vector}

\item{th}{: the threshold to be applied, default is 0}
}
\value{
percentage
}
\description{
Computes the percentage [0,1]of values of a vector that are aboove a given threshold
}
\details{
\code{presenceScore}
}
\author{
Edi Prifti
}
