% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querries.R
\name{profiles2Genebags}
\alias{profiles2Genebags}
\title{profiles2Genebags}
\usage{
profiles2Genebags(profiles)
}
\arguments{
\item{profiles}{: a list of dataframes}
}
\value{
a list of genebags
}
\description{
This function will extract from a list of group profiles a list of identifiers of the matrix.
}
\details{
\code{profiles2Genebags}
}
\author{
Edi Prifti
}
