% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{rdn}
\alias{rdn}
\title{rdn}
\usage{
rdn(x)
}
\arguments{
\item{x}{: data frame or vector}
}
\value{
rounded integer object
}
\description{
fixes an R base issue with rounding
}
\details{
\code{rdn}
}
\author{
Edi Prifti
}

