% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querries.R
\name{selectListSize}
\alias{selectListSize}
\title{selectListSize}
\usage{
selectListSize(l, min.size = 0, max.size = 0, names = FALSE)
}
\arguments{
\item{l}{: a list of vectors or matrixes.}

\item{min.size}{: default = 0 and this has no action. This is the lower threshold for the element's size.}

\item{max.size}{: default = 0 and this has no action. This is the upper threshold for the element's size.}

\item{names}{: default = FALSE the function will return a subset of the list, if TRUE the function will 
return only the names of the elements of the list as identifiers.}
}
\value{
the trimmed subset of the list.
}
\description{
This function will sextract a part of a list based on the length of its components. Typically
a genebag list can be used. This function will work for uniclass lists of vectors and data frames.
}
\details{
\code{selectListSize}
}
\author{
Edi Prifti
}
