% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapreduce.R
\name{splitData}
\alias{splitData}
\title{splitData}
\usage{
splitData(data, shares = 30, pattern = "data_part", rows = TRUE)
}
\arguments{
\item{data}{: the dataframe to be split}

\item{shares}{: the number of shares, default=30}

\item{pattern}{: the name of the split files in the disk preceding the incremental number, default="data_part"}

\item{rows}{: logical parameter, if default rows=="TRUE" the rows will be split}
}
\value{
does not return anything
}
\description{
This function splits a dataframe object on a given number of equally sized shares
and saves them in the disk as RData objects with an incremental name.
}
\details{
\code{splitData}
}
\author{
Edi Prifti
}
