% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapreduce.R
\name{watchProgress}
\alias{watchProgress}
\title{watchProgress}
\usage{
watchProgress(input, output)
}
\arguments{
\item{input}{: a folder containing elements that are to be processed}

\item{output}{: a folder where the processed results will be writen}
}
\value{
nothing
}
\description{
This function prints the progress % of the overall threads that are treated. It is based
on the input and output folders by comparing the number of output objects.
}
\details{
\code{watchProgress}
}
\author{
Edi Prifti
}

