% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapreduce.R
\name{launchTask}
\alias{launchTask}
\title{launchTask}
\usage{
launchTask(input, output, script)
}
\arguments{
\item{input}{: a folder containing elements that are to be processed}

\item{output}{: a folder where the processed results will be writen}

\item{script}{: the R script that is to be executed in parallel}
}
\value{
nothing
}
\description{
This function distributes the calculations as separate processes in a multi-thread server.
}
\details{
\code{launchTask}
}
\note{
the number of processors may be specified. TODO: add the argument
}
\author{
Edi Prifti
}

