% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapreduce.R
\name{mapResults}
\alias{mapResults}
\title{mapResults}
\usage{
mapResults(folder = ".", pattern = "_result.rda", type = "rows")
}
\arguments{
\item{folder}{: the folder where the results are found}

\item{pattern}{: the pattern of the split files in the disk preceding the incremental number, default="data_part"}

\item{type}{: a string c(rows, cols, list) indicating how to merge the data. This should be the same as the one
used in the splitData procedure.}
}
\value{
a merged dataframe : Caution ! The merged results may be shuffeled. It is up to the user to reorder
the data accordingly.
}
\description{
This function loads the results once finished and merges them together in one dataframe
}
\details{
\code{mapResults}
}
\author{
Edi Prifti
}

