% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normalization.R
\name{normFreqTC}
\alias{normFreqTC}
\title{normFreqTC}
\usage{
normFreqTC(dat)
}
\arguments{
\item{dat}{: raw counts data matrix with gene_ids as rownames}
}
\value{
a normalized frequency matrix
}
\description{
converts a raw count matrix onto a frequency matrix using the TC (total count) normalization method.
       This method consists of scaling the signal by the total counts per each sample
}
\details{
\code{normFreqTC}
}
\author{
Edi Prifti
}

