% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotCors2}
\alias{plotCors2}
\title{plotCors2}
\usage{
plotCors2(data, main = "")
}
\arguments{
\item{data}{: a frequency matrix to be visualized}

\item{main}{: the main title of the plot empty by default}
}
\value{
nothing
}
\description{
similar to plotCors but with more levels of colors
}
\details{
\code{plotCors2}
}
\author{
Edi Prifti
}

